+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to the AI::Engine project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2024-09-16
+
+### Added
+
+- Allow passing any parameter when create Assistant.
+
+### Fixed
+
+- Add name to unique index so Assistant migration doesn't fail for being too long.
+
 ## [0.3.0] - 2024-06-27
 
 ### Added